Some compile and warning fixes.

+++ source/wg_eventhandler.cpp
@@ -230,7 +230,7 @@ bool WgEventHandler::IsButtonPressed( int button )
 
 bool WgEventHandler::IsKeyPressed( int native_keycode )
 {
-	for( int i = 0 ; i < m_keysDown.size() ; i++ )
+	for( unsigned int i = 0 ; i < m_keysDown.size() ; i++ )
 		if( native_keycode == m_keysDown[i]->pEvent->NativeKeyCode() )
 			return true;
 
@@ -889,10 +889,10 @@ void WgEventHandler::_processTick( WgEvent::Tick * pEvent )
 
 	// Check if we need to post KEY_REPEAT
 
-	for( int i = 0 ; i < m_keysDown.size() ; i++ )
+	for( unsigned int i = 0 ; i < m_keysDown.size() ; i++ )
 	{
 		KeyDownInfo * pInfo = m_keysDown[i];
-		int64_t timePassed = pInfo->pEvent->Timestamp() - pEvent->Timestamp();
+		int timePassed = (int) (pInfo->pEvent->Timestamp() - pEvent->Timestamp());
 
 		int fraction = 0;
 
@@ -1169,7 +1169,7 @@ void WgEventHandler::_processKeyRepeat( WgEvent::KeyRepeat * pEvent )
 
 	KeyDownInfo * pInfo = 0;
 
-	for( int i = 0 ; i < m_keysDown.size() ; i++ )
+	for( unsigned int i = 0 ; i < m_keysDown.size() ; i++ )
 	{
 		if( pEvent->NativeKeyCode() == m_keysDown[i]->pEvent->NativeKeyCode() )
 		{
@@ -1184,7 +1184,7 @@ void WgEventHandler::_processKeyRepeat( WgEvent::KeyRepeat * pEvent )
 
 	// Post KEY_REPEAT events for gizmos
 
-	for( int i = 0 ; i < pInfo->vGizmos.size() ; i++ )
+	for( unsigned int i = 0 ; i < pInfo->vGizmos.size() ; i++ )
 	{
 		if( pInfo->vGizmos[i] )
 			QueueEvent( new WgEvent::KeyRepeat( pEvent->NativeKeyCode(), pInfo->vGizmos[i].GetRealPtr() ));
@@ -1218,7 +1218,7 @@ void WgEventHandler::_processKeyRelease( WgEvent::KeyRelease * pEvent )
 
 	// Post KEY_RELEASE events for gizmos
 
-	for( int i = 0 ; i < pInfo->vGizmos.size() ; i++ )
+	for( unsigned int i = 0 ; i < pInfo->vGizmos.size() ; i++ )
 	{
 		if( pInfo->vGizmos[i] )
 			QueueEvent( new WgEvent::KeyRelease( pEvent->NativeKeyCode(), pInfo->vGizmos[i].GetRealPtr() ));
